# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ascension / لغتي — Syrian-first English learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscription:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get current student subscription
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSubscription"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create or upgrade subscription (payment integration point)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSubscription"
        "400":
          description: Payment not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SubscriptionInput:
      type: object
      required: [planCode, paymentMethod]
      properties:
        planCode:
          type: string
          enum: [general_english, professional_english]
        paymentMethod:
          type: string
          enum: [sham_cash, cryptocurrency, manual]
        paymentReference:
          type: ["string", "null"]
    StudentSubscription:
      type: object
      required: [id, studentId, planCode, planName, planNameAr, status, startedAt]
      properties:
        id:
          type: number
        studentId:
          type: number
        planCode:
          type: string
        planName:
          type: string
        planNameAr:
          type: string
        status:
          type: string
          enum: [active, pending_payment, expired, cancelled]
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        paymentMethod:
          type: ["string", "null"]
        paymentNote:
          type: ["string", "null"]

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...